Fixed bit-wise operations for originator id;  process ids are now 17-bits and local ids are now 15-bits. This should accomodate OS X > Leopoard. This functionality will change to 32-bits for process ids in an upcoming refactor, with changes to user code to select on originator id and local id.

+++ include/lsst/ctrl/events/CommandEvent.h
@@ -83,12 +83,12 @@ public:
 
     int64_t getOriginatorId();
     short getOriginatorLocalId();
-    short getOriginatorProcessId();
+    int  getOriginatorProcessId();
     int getOriginatorIPId();
 
     int64_t getDestinationId();
     short getDestinationLocalId();
-    short getDestinationProcessId();
+    int  getDestinationProcessId();
     int getDestinationIPId();
 
 protected:
